sample python app/requirements.txt
boto3
langchain
psycopg2
numpy
python-dotenv
flask
langchain_community
pgvector